Ab update publish database system

See merge request minetasia/idarkay/minetasiacore!22

+++ spigot/src/main/java/fr/idarkay/minetasia/core/spigot/commands/money/MoneyGetCommand.java
@@ -41,7 +41,7 @@ public class MoneyGetCommand extends SubCommand implements FixCommand {
             UUID u = plugin.getPlayerUUID(args[1]);
             if(u != null)
             {
-                float m = plugin.getPlayerMoney(u, e);
+                double m = plugin.getPlayerMoney(u, e);
                                                         // player name  amount money type
                 sender.sendMessage(Lang.MONEY_GET.get(lang, Lang.Argument.PLAYER.match(args[1]), Lang.Argument.AMOUNT.match(m), Lang.Argument.MONEY_TYPE.match(e.displayName)));
             }
